requirements.txt
discord
Pillow
selenium